Adding new features

+++ app/Http/Controllers/ArticleController.php
@@ -5,9 +5,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Article;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
+   public function __construct()
+   {
+      $this->middleware("auth")->except(['index', 'detail']);
+   }
+
    public function index()
    {
         $data = Article::latest()->paginate(5);
@@ -44,6 +51,7 @@ class ArticleController extends Controller
       $article->title = request()->title;
       $article->body = request()->body;
       $article->category_id = request()->category_id;
+      $article->user_id = Auth::id();
       $article->save();
 
       return redirect("/articles");
@@ -52,8 +60,35 @@ class ArticleController extends Controller
    public function delete($id)
    {
       $article = Article::find($id);
-      $article->delete();
-
+      
+      if(Gate::allows('delete-article', $article)){
+         $article->delete();
       return redirect("/articles")->with("info", "Deleted an article");
+      }
+      return back()->with('info', 'Unauthorize to delete');
    }
+
+   // I am coding for edit . It is not currently work. 
+  public function update($id){
+
+      $validator = validator(request()->all(),[
+         "title" => "required",
+         "body" => "required",
+         "category_id" => "required",
+      ]);
+
+      if($validator->fails()){
+         return back()->withErrors($validator);
+      }
+
+      $article = Article::find($id);
+      $article->title = request()->title;
+      $article->body = request()->body;
+      $article->category_id = request()->category_id;
+     // $article->user_id = Auth::id();
+      $article->save();
+
+      return redirect("/articles");
+   }
+
 }
