+++ app/Http/Controllers/CommentController.php
@@ -4,9 +4,16 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Comment;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+
+    public function __construct()
+    {
+       $this->middleware("auth"); 
+    }
     public function create(Request $request)
     {
        $request->validate([
@@ -17,6 +24,7 @@ class CommentController extends Controller
        $comment = new Comment;
        $comment->article_id = $request->article_id;
        $comment->content = $request->content;
+       $comment->user_id = Auth::id();
        $comment->save();
 
        return back();
@@ -25,8 +33,12 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
+        
+        if(Gate::allows('delete-comment', $comment)){
+            $comment->delete();
+            return back();
+        }
 
-        return back();
+        return back()->with('info', 'Unauthorize to delete');
     }
 }
